requirements.txt
Django==5.1.2
django-bootstrap5==24.3
django-bootstrap-icons==0.9.0
gunicorn==23.0.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.48
phonenumberslite==8.13.48
